pyproject.toml
[project]
name = "movie-recommendation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chroma>=0.2.0",
    "chromadb>=1.3.4",
    "db>=0.1.1",
    "ipykernel>=7.1.0",
    "numpy>=2.3.4",
    "openai>=2.7.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "rapidfuzz>=3.14.3",
    "requests>=2.32.5",
]
